+++ src/Model/Export/Catalog/ExportPreviewDataProvider.php
@@ -13,17 +13,14 @@ use Omikron\Factfinder\Api\Export\FieldInterface;
 
 class ExportPreviewDataProvider implements DataProviderInterface
 {
-    /**
-     * phpcs:disable Squiz.WhiteSpace.ScopeClosingBrace.ContentBefore
-     * phpcs:disable Squiz.Functions.MultiLineFunctionDeclaration.BraceOnSameLine
-     */
     public function __construct(
         private readonly ExportPreviewProducts $products,
         private readonly ObjectManagerInterface $objectManager,
         private readonly array $productFields,
         private readonly array $entityTypes,
         private readonly array $data,
-    ) {}
+    ) {
+    }
 
     /**
      * @return ExportEntityInterface[]