+++ src/Model/Export/Catalog/ProductType/SimpleDataProvider.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Omikron\Factfinder\Model\Export\Catalog\ProductType;
 
 use Magento\Catalog\Model\Product;
+use Magento\CatalogInventory\Model\Stock\StockItemRepository as StockItem;
 use Omikron\Factfinder\Api\Export\FieldInterface;
 use Omikron\Factfinder\Api\Export\DataProviderInterface;
 use Omikron\Factfinder\Api\Export\ExportEntityInterface;
@@ -12,15 +13,13 @@ use Omikron\Factfinder\Model\Formatter\NumberFormatter;
 
 class SimpleDataProvider implements DataProviderInterface, ExportEntityInterface
 {
-    /**
-     * phpcs:disable Squiz.WhiteSpace.ScopeClosingBrace.ContentBefore
-     * phpcs:disable Squiz.Functions.MultiLineFunctionDeclaration.BraceOnSameLine
-     */
     public function __construct(
         protected Product $product,
         protected NumberFormatter $numberFormatter,
+        protected StockItem $stockItem,
         protected array $productFields = [],
-    ) {}
+    ) {
+    }
 
     /**
      * @inheritdoc
@@ -48,14 +47,14 @@ class SimpleDataProvider implements DataProviderInterface, ExportEntityInterface
             'Short'         => (string) $this->product->getData('short_description'),
             'Deeplink'      => (string) $this->product->getUrlInStore(),
             'Price'         => $this->numberFormatter->format((float) $this->product->getFinalPrice()),
-            'Availability'  => (int) $this->product->isAvailable(),
+            'Availability'  => (int) $this->getAvailability(),
             'HasVariants'   => 0,
             'MagentoId'     => $this->getId(),
         ];
 
         return array_reduce(
             $this->productFields,
-            fn (array $result, FieldInterface $field): array  => [$field->getName() => $field->getValue($this->product)] + $result,
+            fn (array $result, FieldInterface $field): array => [$field->getName() => $field->getValue($this->product)] + $result,
             $data
         );
     }
@@ -63,5 +62,16 @@ class SimpleDataProvider implements DataProviderInterface, ExportEntityInterface
     public function getProduct(): Product
     {
         return $this->product;
+    }
+
+    private function getAvailability(): bool
+    {
+        if ($this->product->hasData('is_salable')) {
+            return $this->product->isAvailable();
+        }
+
+        $quantity = $this->stockItem->get($this->product->getId());
+
+        return $quantity->getIsInStock();
     }
 }