+++ src/Model/Ssr/PriceFormatter.php
@@ -11,15 +11,12 @@ use Omikron\FactFinder\Communication\Version;
 
 class PriceFormatter
 {
-    /**
-     * phpcs:disable Squiz.WhiteSpace.ScopeClosingBrace.ContentBefore
-     * phpcs:disable Squiz.Functions.MultiLineFunctionDeclaration.BraceOnSameLine
-     */
     public function __construct(
         private readonly CommunicationConfig $communicationConfig,
         private readonly PriceCurrencyInterface $priceCurrency,
         private readonly FieldRoles $fieldRoles,
-    ) {}
+    ) {
+    }
 
     public function format(array $searchResult): array
     {