+++ src/Plugin/MapFieldRoles.php
@@ -10,11 +10,9 @@ use Omikron\Factfinder\Model\FieldRoles;
 
 class MapFieldRoles
 {
-    /**
-     * phpcs:disable Squiz.WhiteSpace.ScopeClosingBrace.ContentBefore
-     * phpcs:disable Squiz.Functions.MultiLineFunctionDeclaration.BraceOnSameLine
-     */
-    public function __construct(private readonly CommunicationConfig $communicationConfig) {}
+    public function __construct(private readonly CommunicationConfig $communicationConfig)
+    {
+    }
 
     /**
      * @param FieldRoles $subject