+++ src/Utilities/Validator/ExportPreviewValidator.php
@@ -13,15 +13,12 @@ use Omikron\Factfinder\Exception\ExportPreviewValidationException;
 
 class ExportPreviewValidator implements ValidatorInterface
 {
-    /**
-     * phpcs:disable Squiz.WhiteSpace.ScopeClosingBrace.ContentBefore
-     * phpcs:disable Squiz.Functions.MultiLineFunctionDeclaration.BraceOnSameLine
-     */
     public function __construct(
         private readonly ProductRepositoryInterface $productRepository,
         private readonly ConfigurableType           $configurableType,
         private readonly int                        $entityId,
-    ) {}
+    ) {
+    }
 
     public function validate(): void
     {